Core,Base,BaseGL,OpenGL,Python3: include optimization
* optimized glm and stringconversion includes by splitting up functionality
