+++ include/inviwo/core/io/serialization/serializer.h
@@ -33,7 +33,9 @@
 #include <inviwo/core/io/serialization/serializable.h>
 #include <inviwo/core/util/exception.h>
 #include <inviwo/core/util/stdextensions.h>
-#include <inviwo/core/util/glm.h>
+#include <inviwo/core/util/glmvec.h>
+#include <inviwo/core/util/glmmat.h>
+#include <inviwo/core/util/glmutils.h>
 #include <inviwo/core/io/serialization/serializationexception.h>
 
 #include <flags/flags.h>
