+++ include/inviwo/core/util/glm.h
@@ -31,6 +31,8 @@
 
 #include <inviwo/core/common/inviwocoredefine.h>
 #include <inviwo/core/util/glmvec.h>
+#include <inviwo/core/util/glmmat.h>
+#include <inviwo/core/util/glmutils.h>
 
 #include <warn/push>
 #include <warn/ignore/all>
@@ -73,24 +75,9 @@ static_assert(std::is_trivially_default_constructible<half_float::half>::value, 
 static_assert(std::is_trivial<half_float::half>::value, "");
 static_assert(std::is_pod<half_float::half>::value, "");
 
-using mat2 = glm::mat2;
-using mat3 = glm::mat3;
-using mat4 = glm::mat4;
-using dmat2 = glm::dmat2;
-using dmat3 = glm::dmat3;
-using dmat4 = glm::dmat4;
 using quat = glm::quat;
 
 namespace util {
-
-template <typename T>
-struct is_floating_point : public std::is_floating_point<T> {};
-
-template <>
-struct is_floating_point<half_float::half> : std::true_type {};
-
-template <typename T>
-constexpr bool is_floating_point_v = is_floating_point<T>::value;
 
 template <typename T>
 bool isfinite(const T& v) {
@@ -117,64 +104,6 @@ glm::vec<L, bool, Q> isnan(const glm::vec<L, T, Q>& x) {
 
 template <>
 IVW_CORE_API bool isnan(const half_float::half& v);
-
-template <typename T>
-struct rank : public std::rank<T> {};
-
-template <typename T>
-struct rank<const T> : public rank<T> {};
-
-template <>
-struct rank<half_float::half> : public std::integral_constant<std::size_t, 0> {};
-
-template <glm::length_t L, typename T, glm::qualifier Q>
-struct rank<glm::vec<L, T, Q>> : public std::integral_constant<std::size_t, 1> {};
-
-template <typename T, glm::qualifier Q>
-struct rank<glm::tquat<T, Q>> : public std::integral_constant<std::size_t, 1> {};
-
-template <glm::length_t C, glm::length_t R, typename T, glm::qualifier Q>
-struct rank<glm::mat<C, R, T, Q>> : public std::integral_constant<std::size_t, 2> {};
-
-template <class T>
-constexpr size_t rank_v = rank<T>::value;
-
-template <class T, unsigned N = 0>
-struct extent : std::integral_constant<std::size_t, 1> {};
-
-template <typename T, unsigned N>
-struct extent<const T, N> : public extent<T, N> {};
-
-template <glm::length_t L, typename T, glm::qualifier Q>
-struct extent<glm::vec<L, T, Q>, 0> : public std::integral_constant<std::size_t, L> {};
-
-template <typename T, glm::qualifier Q>
-struct extent<glm::tquat<T, Q>, 0> : public std::integral_constant<std::size_t, 4> {};
-
-template <glm::length_t C, glm::length_t R, typename T, glm::qualifier Q>
-struct extent<glm::mat<C, R, T, Q>, 0> : public std::integral_constant<std::size_t, C> {};
-
-template <glm::length_t C, glm::length_t R, typename T, glm::qualifier Q>
-struct extent<glm::mat<C, R, T, Q>, 1> : public std::integral_constant<std::size_t, R> {};
-
-template <class T, unsigned N = 0>
-constexpr size_t extent_v = extent<T, N>::value;
-
-namespace detail {
-template <typename T, int N>
-struct flat_extent_impl
-    : public std::integral_constant<std::size_t, util::extent<T, N - 1>::value *
-                                                     flat_extent_impl<T, N - 1>::value> {};
-
-template <typename T>
-struct flat_extent_impl<T, 0> : public std::integral_constant<std::size_t, 1> {};
-}  // namespace detail
-
-template <typename T>
-struct flat_extent : detail::flat_extent_impl<T, util::rank<T>::value> {};
-
-template <class T>
-constexpr size_t flat_extent_v = flat_extent<T>::value;
 
 template <class U, class T, class BinaryOperation>
 U accumulate(T x, U init, BinaryOperation op) {
