+++ include/inviwo/core/util/stringconversion.h
@@ -38,7 +38,6 @@
 #include <string_view>
 #include <sstream>
 #include <vector>
-#include <chrono>
 #include <functional>
 #include <cctype>
 #include <locale>
@@ -318,7 +317,6 @@ IVW_CORE_API std::string removeSubString(std::string_view str, std::string_view 
 IVW_CORE_API std::string removeFromString(std::string str, char char_to_remove = ' ');
 IVW_CORE_API void replaceInString(std::string& str, std::string_view oldStr,
                                   std::string_view newStr);
-IVW_CORE_API std::string parseTypeIdName(std::string str);
 
 IVW_CORE_API std::string toUpper(std::string_view str);
 IVW_CORE_API std::string toLower(std::string_view str);
@@ -356,37 +354,5 @@ IVW_CORE_API bool iCaseLess(std::string_view l, std::string_view r);
 struct IVW_CORE_API CaseInsensitiveCompare {
     bool operator()(std::string_view a, std::string_view b) const;
 };
-
-/**
- * \brief convert the given duration from milliseconds to a string.
- * The returned string will have the format "%dd %dh %dmin %dsec %.3fms", where days, hours,
- * minutes, seconds, ... are suppressed up to the first non-zero unit if not needed. Milliseconds
- * and seconds are combined if larger than 1 second.
- *
- * @param ms  in milliseconds
- * @param includeZeros   if true, time units for zero values are always shown, e.g.
- *                       "2d 0h 0min 23s" vs. "2d 23s" and "2h 0min 0s" vs. "2h"
- * @param spacing   if true, a space is inserted between digits and units
- * @return duration formatted as string
- */
-IVW_CORE_API std::string msToString(double ms, bool includeZeros = true, bool spacing = false);
-
-/**
- * \brief convenience function for converting a std::chrono::duration to a string calling
- * msToString(double).
- *
- * @param duration       duration
- * @param includeZeros   if true, time units for zero values are always shown, e.g.
- *                       "2d 0h 0min 23s" vs. "2d 23s" and "2h 0min 0s" vs. "2h"
- * @param spacing   if true, a space is inserted between digits and units
- * @return duration formatted as string
- */
-template <class Rep, class Period = std::ratio<1>>
-std::string durationToString(std::chrono::duration<Rep, Period> duration, bool includeZeros = true,
-                             bool spacing = false) {
-    using milliseconds = std::chrono::duration<double, std::milli>;
-    auto ms = std::chrono::duration_cast<milliseconds>(duration);
-    return msToString(ms.count(), includeZeros, spacing);
-}
 
 }  // namespace inviwo
