+++ modules/opengl/src/inviwoopengl.cpp
@@ -29,7 +29,12 @@
 
 #include <modules/opengl/inviwoopengl.h>
 #include <inviwo/core/util/logcentral.h>
+#include <inviwo/core/util/typeid.h>
 #include <sstream>
+
+#if defined(_WIN32)
+#include <inviwo/core/util/stringconversion.h>
+#endif
 
 namespace inviwo {
 
@@ -106,9 +111,9 @@ void LogGLError(const std::type_info& source, std::string_view fileName,
     // There might be several errors, call glGetError in a loop:
     // https://www.opengl.org/sdk/docs/man2/xhtml/glGetError.xml
     while ((err = glGetError()) != GL_NO_ERROR && maxErrors--) {
-        LogCentral::getPtr()->log(parseTypeIdName(std::string(source.name())), LogLevel::Error,
-                                  LogAudience::Developer, fileName, functionName, lineNumber,
-                                  getGLErrorString(err));
+        LogCentral::getPtr()->log(util::parseTypeIdName(std::string(source.name())),
+                                  LogLevel::Error, LogAudience::Developer, fileName, functionName,
+                                  lineNumber, getGLErrorString(err));
     }
 }
