+++ src/core/util/typeid.cpp
@@ -2,7 +2,7 @@
  *
  * Inviwo - Interactive Visualization Workshop
  *
- * Copyright (c) 2020-2021 Inviwo Foundation
+ * Copyright (c) 2022 Inviwo Foundation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,54 +27,36 @@
  *
  *********************************************************************************/
 
-#pragma once
-
-#include <warn/push>
-#include <warn/ignore/all>
-
-#include <glm/vec2.hpp>
-#include <glm/vec3.hpp>
-#include <glm/vec4.hpp>
-
-#include <warn/pop>
-
-// extracted from <glm/gtx/std_based_type.hpp> to avoid including glm.hpp
-namespace glm {
-typedef vec<1, std::size_t, defaultp> size1;
-typedef vec<2, std::size_t, defaultp> size2;
-typedef vec<3, std::size_t, defaultp> size3;
-typedef vec<4, std::size_t, defaultp> size4;
-typedef vec<1, std::size_t, defaultp> size1_t;
-typedef vec<2, std::size_t, defaultp> size2_t;
-typedef vec<3, std::size_t, defaultp> size3_t;
-typedef vec<4, std::size_t, defaultp> size4_t;
-}  // namespace glm
+#include <inviwo/core/util/typeid.h>
+#include <inviwo/core/util/stringconversion.h>
 
 namespace inviwo {
 
-using ivec2 = glm::ivec2;
-using ivec3 = glm::ivec3;
-using ivec4 = glm::ivec4;
-using vec2 = glm::vec2;
-using vec3 = glm::vec3;
-using vec4 = glm::vec4;
-using dvec2 = glm::dvec2;
-using dvec3 = glm::dvec3;
-using dvec4 = glm::dvec4;
-using bvec2 = glm::bvec2;
-using bvec3 = glm::bvec3;
-using bvec4 = glm::bvec4;
-using uvec2 = glm::uvec2;
-using uvec3 = glm::uvec3;
-using uvec4 = glm::uvec4;
-using size2_t = glm::size2_t;
-using size3_t = glm::size3_t;
-using size4_t = glm::size4_t;
-using i64vec2 = glm::i64vec2;
-using i64vec3 = glm::i64vec3;
-using i64vec4 = glm::i64vec4;
-using u64vec2 = glm::u64vec2;
-using u64vec3 = glm::u64vec3;
-using u64vec4 = glm::u64vec4;
+namespace util {
+
+std::string parseTypeIdName(std::string str) {
+
+#if defined(__clang__) || defined(__GNUC__)
+    struct handle {
+        char* p;
+        handle(char* ptr) : p(ptr) {}
+        ~handle() { std::free(p); }
+    };
+    const char* cstr = str.c_str();
+    int status = -4;
+    handle result(abi::__cxa_demangle(cstr, nullptr, nullptr, &status));
+    if (status == 0) str = result.p;
+#else
+    replaceInString(str, "class", "");
+    replaceInString(str, "const", "");
+#if defined(_WIN64) || defined(__x86_64__) || defined(__ppc64)
+    replaceInString(str, "__ptr64", "");
+#endif
+#endif
+    replaceInString(str, "inviwo::", "");
+    return removeFromString(removeFromString(str, '*'), ' ');
+}
+
+}  // namespace util
 
 }  // namespace inviwo
